verifica si la orden tiene una factura asociada antes de permitir su eliminación

+++ apps/ventas/views.py
@@ -5,6 +5,7 @@ from django.contrib import messages
 from django.core.paginator import Paginator
 from django.db.models import Q
 from django.db import connection
+from django.db import IntegrityError
 
 def lista_ventas(request):
     busqueda = request.GET.get("buscar", "")
@@ -34,7 +35,9 @@ def crear_venta(request):
                                              COALESCE(MAX(id_orden), 1))
                                FROM orden;
                                """)
-            formulario.save()
+            orden = formulario.save(commit=False) #por el momento no guardar
+            orden.total = 0 #total está en 0 al inicio
+            orden.save()
             messages.success(request, "Orden registrada correctamente.")
             return redirect("lista_ventas")
     else:
@@ -61,8 +64,19 @@ def editar_venta(request, id_orden):
 
 def eliminar_venta(request, id_orden):
     orden = get_object_or_404(Orden, pk=id_orden)
+
     if request.method == "POST":
-        orden.delete()
-        messages.success(request, "Orden eliminada correctamente.")
+        if orden.estado == "Entregada":
+            messages.error(request, "Restricción de Auditoría: No se puede eliminar una orden en estado 'Entregada'.")
+            return redirect("lista_ventas")
+
+        try:
+            orden.delete()
+            messages.success(request, "Orden eliminada correctamente.")
+        except IntegrityError:
+            # Captura si la orden ya tiene una FACTURA amarrada
+            messages.error(request, "No es posible eliminar esta orden porque ya tiene una Factura Electrónica emitida en el histórico.")
         return redirect("lista_ventas")
+
+    # Si entra por GET (solo dando clic al botón inicial), muestra la pantalla de confirmación normal
     return render(request, "ventas/eliminar.html", {"orden": orden})
